// src/features/media-blocks/schemas/media-blocks.ts
export const videoGalleryItemSchema = z.object({
  id: z.string(),
  videoUrl: z.string().default(""),
  videoMediaAssetId: z.string().default(""),
  embedUrl: z.string().default(""),
  thumbnailUrl: z.string().default(""),
  thumbnailMediaAssetId: z.string().default(""),
  titleEn: z.string().default(""),
  titleAr: z.string().default(""),
  categoryEn: z.string().default(""),
  categoryAr: z.string().default(""),
  playlistId: z.string().default(""),
});
// ...
export type VideoGalleryItem = z.infer<typeof videoGalleryItemSchema>;